yalla go

+++ PostingFile.py
@@ -5,9 +5,10 @@ import Term
 class PostingFile:
     MAX_ROWS_TO_READ = 3
 
-    def __init__(self, path, name):
+    def __init__(self, path, name, is_delete=True):
         self.path = path
         self.name = name
+        self.is_delete = is_delete
         self.file = open(path, 'r')
         self.term_name = None
         self.postings = None
@@ -21,7 +22,8 @@ class PostingFile:
         line = self.lines_array[0]
         if line.strip() == '':
             self.file.close()
-            os.remove(self.path)
+            if self.is_delete is True:
+                os.remove(self.path)
             return False
         self.parse_posting_file_line(line)
         del self.lines_array[0]
@@ -43,5 +45,17 @@ class PostingFile:
         postings = line[cut_point:]
         postings_size = len(postings)
         # remove the ':[' and ']\n'
-        postings = postings[2:postings_size - 2]
+        chars_to_del = 2 if postings[1] == '[' else 1
+        postings = postings[chars_to_del:postings_size - chars_to_del]
         self.postings = Term.postings_string_to_array_of_terms(self.term_name, postings)
+
+    def read_line_by_position(self, position, is_cache):
+        self.file.seek(int(position))
+        if not is_cache:
+            self.parse_posting_file_line(self.file.readline())
+        else:
+            pass
+        # TODO add something to read & parse the right position if it's a cache record
+
+    def close_posting_file(self):
+        self.file.close()
